# Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a RESTful API built with Node.js(Typescript), Nginx, MongoDB, RabbitMQ and Minio(object store with S3ish sdk).",
    "version": "1.0.0",
    "title": "Swagger Petstore"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/image": {
      "post": {
        "tags": [
          "image"
        ],
        "summary": "Upload an image to the service",
        "description": "",
        "operationId": "uploadImage",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "upload",
            "type": "file",
            "description": "Image file to be uploaded",
            "required": true
          },
          {
            "in": "header",
            "name": "x-username",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Image saved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Something Broke :( "
          }
        }
      },
      "get": {
        "tags": [
          "image"
        ],
        "summary": "Get an image from the service",
        "description": "",
        "operationId": "getImage",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "image/png",
          "image/gif",
          "image/jpeg",
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "type": "string",
            "description": "File name to get",
            "required": true
          },
          {
            "in": "header",
            "name": "x-username",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Image file"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No image found"
          },
          "500": {
            "description": "Something Broke :( "
          }
        }
      },
      "delete": {
        "tags": [
          "image"
        ],
        "summary": "Delete an image from the service",
        "description": "",
        "operationId": "deleteImage",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "type": "string",
            "description": "File name to delete",
            "required": true
          },
          {
            "in": "header",
            "name": "x-username",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Image deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No image found"
          },
          "500": {
            "description": "Something Broke :( "
          }
        }
      }
    },
    "/imagenames": {
      "get": {
        "tags": [
          "imagenames"
        ],
        "summary": "Get all Image names associated with the current user",
        "description": "",
        "operationId": "getImageNames",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-username",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Image saved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No image names found"
          },
          "500": {
            "description": "Something Broke :( "
          }
        }
      }
    },
    "/usage": {
      "get": {
        "tags": [
          "usage"
        ],
        "summary": "Get a realtime usage report",
        "description": "",
        "operationId": "getUsage",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-username",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Image saved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Something Broke :( "
          }
        }
      }
    }
  },
  "externalDocs": {
    "description": "Github",
    "url": "https://github.com/sefij/image-api"
  }
}